src/gaming_scheduler_bot/bot.py
import discord
from discord.ext import commands

import os
import logging


token = os.getenv("DISCORD_TOKEN")

handler = logging.FileHandler(filename="bot.log", encoding="utf-8", mode="w")

intents = discord.Intents.default()
intents.message_content = True

bot = commands.Bot(command_prefix="!", intents=intents)


@bot.event
async def on_ready():
    print(f"{bot.user.name} is ready!")


def run():
    bot.run(token, log_handler=handler, log_level=logging.DEBUG)
