pyproject.toml
[project]
name = "gaming-scheduler-bot"
version = "0.1.0"
description = "Group gaming session scheduler bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
